@@ -6,7 +6,7 @@ homepage: https://github.com/Fliggy-Mobile/fbroadcast
 email: coorchice.cb@alibaba-inc.com
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.2 <3.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
